push modifiche

+++ siw-product/src/main/java/it/uniroma3/siw/service/CommentService.java
@@ -1,11 +1,14 @@
 package it.uniroma3.siw.service;
 
 import it.uniroma3.siw.model.Comment;
+import it.uniroma3.siw.model.Product;
+import it.uniroma3.siw.model.User;
 import it.uniroma3.siw.repository.CommentRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.time.LocalDateTime;
 import java.util.List;
 
 @Service
@@ -30,6 +33,18 @@ public class CommentService {
         return commentRepository.findByUserId(userId);
     }
     
+    
+    public Comment findByProductAndUser(Product product, User user) {
+    	return commentRepository.findByProductAndUser(product, user);
+    }
+    
+    
+    public boolean existsByProductAndUser(Product product, User user) {
+    	return commentRepository.existsByProductAndUser(product, user);
+    }
+    
+    
+    
     @Transactional
     public Comment save(Comment comment) {
         return commentRepository.save(comment);
@@ -40,6 +55,7 @@ public class CommentService {
         Comment comment = commentRepository.findById(id).orElse(null);
         if (comment != null) {
             comment.setText(newText);
+            comment.setCreatedAt( LocalDateTime.now() );
             return commentRepository.save(comment);
         }
         return null;
@@ -47,6 +63,12 @@ public class CommentService {
     
     @Transactional
     public void deleteById(Long id) {
+    	
+    	if(!this.commentRepository.existsById(id))
+			return;
         commentRepository.deleteById(id);
     }
+    
+    
+    
 }
